app3/requirements.txt
flask
#boto has instrumentation but not boto3
botocore==1.21

gunicorn==20.1.0


opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-propagator-b3
opentelemetry.exporter.zipkin
opentelemetry-instrumentation-flask
opentelemetry.instrumentation.botocore
opentelemetry-instrumentation-logging